+++ src/components/NotificationQueue/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Style from './style';
+
+const NotificationQueue = (): JSX.Element => <Style className='notificationQueue'></Style>;
+
+export default NotificationQueue;